refactor(symengine): split seFromSE.c along its three real seams
habit-hooks flagged four oversized/high-complexity functions and the file
itself.  The seams were real, so this is a redistribution rather than a
line-count trim:

- src/seFromSEarena.h  -- hold every string the emitter builds, so the walk
  can return const char * with no ownership bookkeeping and never touch R's
  allocator.
- src/seFromSEnames.h  -- the name and constant vocabulary shared by symengine
  and rxode2: undo the mangling, render a number as R's parser plus
  as.character() would, and recognize text that stands for a C constant.
- src/seFromSE.c       -- walk a parse tree and emit rxode2 text.

Two missing abstractions came out of it:

- seNamedConstant().  .rxFromSE() asks "is this finished string a named C
  constant?" in three separate if-chains -- one in the binary-operator branch,
  one at the end of the generic call branch, one inside log().  They are the
  same question, so they are now one table.  That is also what let the
  M_SQRT_PI list go missing when only two of the three sites were ported.
- seIsWrapper()/seIsNumberNode()/seIsLeafNode().  The grammar spells the
  precedence ladder out, so seIsBareNumber(), seStripP() and seFold() each
  repeated the list of single-child pass-through productions.

Also extracted seEmitPower(), seEmitLog(), seCallName(), seEmitSymbol(),
seFoldLeaf(), seFoldBinary(), and the seThEtaName()/seOrdinalEnd()/seLastDy()
pieces the three mangled-name parsers were each spelling out.

The fixture generator's model corpus moves to
inst/tools/symengineFixtureCorpus.R: growing the corpus and changing the
capture procedure are different jobs.  The generator reproduces a
byte-identical fixture.

No behavior change: still 434 of 592 accepted with zero byte differences, and
the 1336 assertions in the symbolic suites still pass.
